refactor(compilation): moved compilation classes to compilation package
